# html/app/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bridge API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://your-api-url.com/api/v1"
        }
    ],
    "paths": {
        "/api/v1": {},
        "/roles": {
            "post": {
                "tags": [
                    "Role"
                ],
                "operationId": "af6b32baf16ebe819bcc525bcd8692ff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRoleParams"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "create",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UidSuccessApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKey": []
                    }
                ]
            }
        },
        "/roles/{id}": {
            "get": {
                "tags": [
                    "Role"
                ],
                "operationId": "e67fb0b467d6552f1fc088fade9fe9dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleSuccessApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "no rows found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKey": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Uid": {
                "properties": {
                    "uid": {
                        "type": "string",
                        "example": "01JC083YJ6K9D91BW4WQQBWWZH",
                        "nullable": false
                    }
                },
                "type": "object"
            },
            "Role": {
                "properties": {
                    "uid": {
                        "type": "string",
                        "example": "01JC083YJ6K9D91BW4WQQBWWZH",
                        "nullable": false
                    },
                    "role_name": {
                        "type": "string",
                        "example": "mock string",
                        "nullable": false
                    },
                    "role_name_slug": {
                        "type": "string",
                        "example": "mock-string",
                        "nullable": false
                    },
                    "description": {
                        "type": "string",
                        "example": "mock description",
                        "nullable": false
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2024-11-06T08:05:38.486431+00:00",
                        "nullable": false
                    },
                    "created_by": {
                        "type": "string",
                        "example": "01JC083YJ6K9D91BW4WQQBWWZH",
                        "nullable": false
                    },
                    "is_activated": {
                        "type": "boolean",
                        "example": true,
                        "nullable": false
                    },
                    "is_activated_updated_at": {
                        "type": "string",
                        "example": "2024-11-06T08:05:38.486431+00:00",
                        "nullable": false
                    },
                    "is_activated_updated_by": {
                        "type": "string",
                        "example": "01JC083YJ6K9D91BW4WQQBWWZH",
                        "nullable": false
                    }
                },
                "type": "object"
            },
            "CreateRoleParams": {
                "properties": {
                    "role_name": {
                        "type": "string",
                        "example": "mock string",
                        "nullable": false
                    },
                    "description": {
                        "type": "string",
                        "example": "mock description",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UidSuccessApiResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 200,
                        "nullable": false
                    },
                    "message": {
                        "type": "string",
                        "example": "OK",
                        "nullable": false
                    },
                    "timestamp": {
                        "type": "string",
                        "example": "2024-11-06T08:05:38.486431+00:00",
                        "nullable": false
                    },
                    "data": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Uid"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "RoleSuccessApiResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 200,
                        "nullable": false
                    },
                    "message": {
                        "type": "string",
                        "example": "OK",
                        "nullable": false
                    },
                    "timestamp": {
                        "type": "string",
                        "example": "2024-11-06T08:05:38.486431+00:00",
                        "nullable": false
                    },
                    "data": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Role"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "SuccessResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 200,
                        "nullable": false
                    },
                    "message": {
                        "type": "string",
                        "example": "OK",
                        "nullable": false
                    },
                    "timestamp": {
                        "type": "string",
                        "example": "2024-11-06T08:05:38.486431+00:00",
                        "nullable": false
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 400,
                        "nullable": false
                    },
                    "message": {
                        "type": "string",
                        "example": "mock error message",
                        "nullable": false
                    },
                    "timestamp": {
                        "type": "string",
                        "example": "2024-11-06T08:05:38.486431+00:00",
                        "nullable": false
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiKey": {
                "type": "apiKey",
                "name": "x-api-key",
                "in": "header"
            },
            "Token": {
                "type": "apiKey",
                "name": "x-token",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Role",
            "description": "Role"
        }
    ]
}